Reduce TraceOperation size by changing BranchProbability from double to float
This commit reduces TraceOperation size from 96 bytes to 64 bytes (33% reduction)
by changing BranchProbability from double (8 bytes) to float (4 bytes).

Changes:
- Changed BranchProbability type alias from double to float in TraceOperation.hpp
- Changed TrueProbability type alias from double to float in val_bool.hpp
- Updated addCmpOperation signature to accept float instead of double
- Updated ExecutionTrace implementation to match new signature

Memory impact:
- InputVariant: 16 bytes → 8 bytes (50% reduction)
- BoundedInputArray<InputVariant,4>: 72 bytes → 36 bytes (50% reduction)
- TraceOperation: 96 bytes → 64 bytes (33% reduction)
- For 10,000 operations: 960 KB → 640 KB (33% memory savings)

The float precision (6-7 significant digits) is sufficient for branch probability
hints (0.0 to 1.0), as these are optimization hints rather than precise values.
This change improves memory efficiency and cache locality during tracing with
no functional impact on code generation quality.

All 51 tests pass successfully.

+++ nautilus/src/nautilus/tracing/ExecutionTrace.cpp
@@ -186,7 +186,7 @@ TypedValueRef& ExecutionTrace::addOperationWithResult(Snapshot& snapshot, Op& op
 
 // Adds a comparison operation to the execution trace
 // This consists of a snapshot, the comparison input, two blocks for true and false branches, and the branch probability
-void ExecutionTrace::addCmpOperation(Snapshot& snapshot, const TypedValueRef& condition, const double probability) {
+void ExecutionTrace::addCmpOperation(Snapshot& snapshot, const TypedValueRef& condition, const float probability) {
 	if (blocks.empty()) {
 		createBlock();
 	}
